Move transaction error code into new module
